refactor(api): split validateUserProfileData from user service
